No more unicode spaces! Formatting looks solid. Also fixed one typo in card text.

+++ GtR/GtR/MeasuredTextFragment.cs
@@ -3,10 +3,10 @@
     public class MeasuredTextFragment
     {
         public TextFragment TextFragment { get; }
-        public int Width { get; }
+        public float Width { get; }
         public int Height { get; }
 
-        public MeasuredTextFragment(int width, int height, TextFragment textFragment)
+        public MeasuredTextFragment(float width, int height, TextFragment textFragment)
         {
             Width = width;
             Height = height;
